+++ heatingassistant/app/static/js/pages/configuration.js
@@ -5,7 +5,7 @@ import { renderRoomList } from '../config/config-room-list.js?v=124';
 import { renderRoomEditor } from '../config/config-room-editor.js?v=124';
 import { renderSourceList } from '../config/config-source-list.js?v=124';
 import { renderSourceEditor } from '../config/config-source-editor.js?v=133';
-import { renderSystem, renderSystemParams } from '../config/config-system.js?v=124';
+import { renderSystem, renderSystemParams } from '../config/config-system.js?v=151';
 import { renderAdvanced } from '../config/config-advanced.js?v=134';
 
 export function renderConfiguration(container, rooms, state, connection, hass, slug) {